src: add Embeddings for llama
This commit adds the ability to generate embeddings using the llama.

The motivation for this is to be able to use llama for embeddings in
combination with a vector store, like Qdrant.

This commit also adds an example that demonstrates how to use the
llm-chain-llama crate for generating embeddings and then use
the Qdrant vector store for storing and searching for similar
documents.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>