llama: clear kv cache for each run_model call
This commit adds a call to `llama_kv_cache_clear` for each call to
`run_model`. This is done because the same sequence id is currently
being used for each call to `run_model` which can cause tokens from a
previous call to be in the catch. This can cause the model to use tokens
from a previous decode call in the attention mechanism which can cause
the model to generate incorrect information.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>