llama: replace cuda feature with env var
This commit removes the `cuda` feature and replaces it with an
environment variable `LLM_CHAIN_CUDA`.It also adds an additional
environment variable `LLM_CHAIN_CUDA_LIB_PATH` which should be set to
the path of the CUDA library directory, and include link libraries
required for CUDA. Finally `-fPIC` is added to `CMAKE_CUDA_FLAGS`.

The motivation for this is that having the feature `cuda` was not
working as this will cause the build to fail if infact the CUDA
library is not installed. I think that this may not have been the case
with earlier versions of llama.cpp and perhaps why this was not an
issue before.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>